apps/streamlit/requirements.txt
streamlit==1.37.1
snowflake-snowpark-python==1.18.0
snowflake-connector-python==3.12.2
pydantic==2.8.2
sqlparse==0.5.1
